Completed the queue-deletion and added an extra test to ensure that no memory is lost when a queue is deleted.
There are still problems regarding the locking of a node before it is
deleted; when there is only a single node in the queue it is possible
for a caller to access head while tail is locked and being deleted
(since both head and tail point to the same node).

This is not such a problem as head is onyl used for adding nodes to the
queue while tail is only used for removing nodes from a queue.