FEATURE: return lifetime of element in a queue.
When an element is removed, a struct timspec is returned indicating the
duration of the elements time in the queue.

This was added so callers (like amq) can accurately report queue
waiting-times to the user (or their callers). The caller (or the user)
can then adjust their behaviour accordingly, and/or report on the
durations.