BUGFIX: deadlocking was possible when using two mutexes.
Changed cmq implementation to lock entire queue with a single mutex
instead of trying to lock the head and tail independently. Locking the
head and tail meant that sometimes, when both are pointing to the same
element, they can deadlock because of the order that they acquire locks.

With only a single mutex there is no possibility of a deadlock, but it
does mean that the entire queue is locked for longer periods than when
using two mutexes (although this is mitigated somewhat by the fact that
locking happens less often).

The correct way to do this would be using 'trylock' and some sort of
random nanosleep (or pthread_yield()): any thread attempting to acquire
a lock will have to do so in a specific order and relinquish all locks
if any of the attempts to lock fail.